Create new AMQP properties for each publication
The timestamp was only getting set on the first time getBasicProperties()
was called.

Since there are no setter methods on the AMQP.BasicProperties class, we
can't just set a new timestamp on the existing object every time. And it's
ugly to reconstruct a new one based off the existing one.

Change the method's logic to only initialise the headers on the first call,
and then return a new instance of the AMQP.BasicProperties object on each
subsequent call.

Change-Id: I7436bc5ffadc566d30d7ae465fcf618d1e983653