AMQProperties: Fix IllegalArgumentException when setting properties
IllegalArgumentException is raised:

  Attempt to get java.lang.Integer field "<name>" with illegal data
  type conversion to int

this can be reproduced for example when a branch is created and the
ref-updated message is emitted. The result is an internal server error
returned to the REST call, although the branch was created successfully.

Instead of calling getInt() and friends, use the valueOf() methods on
the boxing classes with toString() on the value returned from the basic
get() method.

Change-Id: Iba95aba743bb76eeef5a193950de4f1b1f7c3e71