Update patch set 1
Patch Set 1:

> I have a feeling that this was exacerbated by the changes that I made in I1889cd4da702ce43a4e087f21c27680900abd5a9
> Before that change, the errors would have been silently ignored.

I'm not sure I follow you here.
All your change did was change log level from info to warn.
The change is not in the rabbitmq plugin we run in our production environment, yet we have the same log spew [1]
The AMQProperties constructor is only called once and I can't see the correlation between branch creation and this call.
This code sets the headers whereas the ref-updated message attribute is gerrit-event-specific, is it not.

[2015-05-20 11:23:25,955] INFO  com.googlesource.gerrit.plugins.rabbitmq.config.AMQProperties : Attempt to get java.lang.Integer field "com.googlesource.gerrit.plugins.rabbitmq.config.section.Gerrit.port" with illegal data type conversion to int

Patch-set: 1