Update patch set 1
Patch Set 1:

> I'm not sure I follow you here.
 > All your change did was change log level from info to warn.

It also changed the exception handling to only catch IllegalAccessException instead of the more generic Exception.

 > The change is not in the rabbitmq plugin we run in our production
 > environment, yet we have the same log spew [1]

In the version without my change it will just log the exception.  After my change NullPointerException and IllegalArgumentException are not caught and will be propagated up which results in the 500 error.

 > The AMQProperties constructor is only called once and I can't see
 > the correlation between branch creation and this call.

Could just be that in my tests I was not doing anything other than trying to create branches via the REST API.

Patch-set: 1