MessagePublisher: Threads cannot be restarted
If publisherThread is not alive when a new event arrives we currently
try to restart the dead Thread which results in a
IllegalThreadStateException.
If publisherThread has died, create a new Thread and start it.

Synchronize the start-new-thread logic to avoid race conditions
that potentially could start new threads unnecessarily and leave
existing, live, threads orphaned.

Change-Id: I772a3275b16be7f2fb31c141ad9a71f162d53ce1