MessagePublisher: Threads cannot be restarted
If publisherThread dies we would try to restart the same Thread on
line 68, which results in an IllegalThreadStateException.
If publisherThread has died, create a new Thread and start it.

Synchronize the start-new-thread logic to avoid race conditions
that potentially could start new threads unnecessarily and leave
existing, live, threads orphaned.

Change-Id: I772a3275b16be7f2fb31c141ad9a71f162d53ce1